% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_mir_mrna_exprs.R
\name{combine_mir_mrna_exprs}
\alias{combine_mir_mrna_exprs}
\title{Combine miRNA and mRNA expression data}
\usage{
combine_mir_mrna_exprs(mirna_data, mrna_data)
}
\arguments{
\item{mirna_data}{a numeric matrix or data frame containing the miRNA expression profile.
Rows are miRNA names and columns are samples.}

\item{mrna_data}{a numeric matrix or data frame containing the mRNA gene expression profile.
Rows are gene names and columns are samples.}
}
\value{
Returns an object of class MirMrnaExprs which is a data.frame containing the combined gene expression data.
It includes also two attributes "mirna_index" and "mrna_index". Rows are samples and columns are miRNA/mRNA names.
}
\description{
Combine miRNA and mRNA expression data.
The output dataset will be used for statistical approach-based miRNA target predictions.
The name and the number of samples in the two data set should be identical.
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
A. Kassambara. Statistical Tools for High-throughput Data Analysis. http://www.sthda.com
}

