% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_mir_pvt.R
\name{combine_mir_pvt}
\alias{combine_mir_pvt}
\title{Combine miRNA predicted and validated targets}
\usage{
combine_mir_pvt(p_targets, v_targets)
}
\arguments{
\item{p_targets, v_targets}{an object of class MirTarget}
}
\value{
Returns an object of class MirTarget which is a data frame containing
 combined predicted and validated targets for each miRNA. The final results
 might contains different status of miRNA-target interactions (MTIs):
 \itemize{ \item V: for validated MTIs \item P: for predicted MTIs}
}
\description{
A helper function for combining miRNA predicted and validated
 targets. Duplicated miRNA-target interactions are automatically removed.
}
\examples{
\dontrun{
# Example of miRNAs
miRNAs <- c("hsa-miR-16", "hsa-miR-155")

# Validated target from miRecords
v_targets <- get_mirecords_vt(miRNAs)

# Get predicted target
p_targets <- get_mirecords_pt(miRNAs)

# Combine
targets <- combine_mir_pvt(p_targets, v_targets)
head(targets)
}

}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
A. Kassambara. Statistical Tools for High-throughput Data
 Analysis. http://www.sthda.com
}

