% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_mir_vt.R
\name{combine_mir_vt}
\alias{combine_mir_vt}
\title{Combine miRNA validated targets from differents sources (miRecords and miRTarbase)}
\usage{
combine_mir_vt(...)
}
\arguments{
\item{...}{a list of MirTarget object to combine. MirTarget object can be
generated by the functions get_mirecords_*() or by get_mirtarbase_*().}
}
\value{
Returns an
  object of class MirTarget which is a data frame containing combined validated
  targets for each miRNA.
}
\description{
A helper function for combining miRNA targets from differents
 sources. Duplicated miRNA-target interactions are automatically removed.
}
\examples{
\donttest{
# Example of miRNAs
miRNAs <- c("hsa-miR-16", "hsa-miR-155")

# Validated target from miRecords
mirecords_vt <- get_mirecords_vt(miRNAs)

# Validated target from miRTarbase
mirtarbase_vt <- get_mirtarbase_vt(miRNAs)

v_targets <- combine_mir_vt(mirecords_vt, mirtarbase_vt)
head(v_targets)
}

}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
A. Kassambara. Statistical Tools for High-throughput Data
 Analysis. http://www.sthda.com
}

