% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{corr_mir_mrna}
\alias{correlation}
\alias{viz_mrna_per_mir}
\alias{viz_neg_cor_fraction}
\title{Correlation between miRNA and mRNA expression}
\usage{
corr_mir_mrna(mir_targets, mir_mrna_exprs, mirs_of_interest = "all",
  method = "pearson", stand = TRUE, sortby = c("cor.coeff", "p.value",
  "fdr", "none"), show_progress = TRUE)

viz_mrna_per_mir(object, coeff = 0.5, lab.size = 8, lab.angle = 45,
  output = c("plot", "table"))

viz_neg_cor_fraction(object, output = c("plot", "table"))
}
\arguments{
\item{mir_targets}{an object of class MirTarget containing the miRNA-mRNA
target interactions. It must contains three columns: mirna_name |
target.gene_name | status.}

\item{mir_mrna_exprs}{an object of class MirMrnaExprs, as generated by the
function combine_mir_mrna_exprs(). It must contains the expression of
miRNAs and mRNAs.}

\item{mirs_of_interest}{a vector containing the names of miRNA for which you
want to compute correlation analysis against target genes. Default value is
"all".}

\item{method}{method to be used for correlation analyses. Allowed values are
one of "pearson", "kendall", or "spearman.}

\item{stand}{logical value. If TRUE the miR/mRNA expression data are
standardized before computing correlation. Default value is TRUE.}

\item{sortby}{Specify how to sort the result. Allowed values are one of
"cor.coeff", "p.value", "fdr", or "none": Abbreviation is allowed. Default
value is "cor.coeff".}

\item{show_progress}{logical value. If TRUE, a progress bar is shown.}

\item{object}{an object of class CorMirTarget as generated using}

\item{coeff}{correlation coefficient threshold. miRNAs having absolute}

\item{lab.size}{size of x axis labels}

\item{lab.angle}{rotation angle of x axis labels
correlation >= coeff with one or more mRNAs are kept.}

\item{output}{allowed values are "plot" or "table"}
}
\value{
Returns an object of class CorMirTargets which is a data.frame
  containing correlation between miRNAs and their mRNAs. Column names of the results are as follow:
  \itemize{
  \item \bold{mirna_name}: Name of miRNAs
  \item \bold{target.gene_name}: The name of target genes
  \item \bold{cor.coeff}: Correlation coefficient
  \item \bold{p.value}: The p-value of the correlation test
  \item \bold{fdr}: Adjusted p-value using BH multiple testing correlation.
  This calculated separatly for the targets of each miRNA.
  \item \bold{status}: miRNA Target status. P for predicted, V for validated; PV for predicted and validated.
  }
}
\description{
Compute the correlation between the expression of each miRNA and
  its potential mRNA targets. If a gene is targeted by a miRNA we expect that
  the expression profile of the gene and the miRNA are anticorrelated.
}
\section{Functions}{
\itemize{
\item \code{corr_mir_mrna}: Compute correlation between miRNA and mRNA expression

\item \code{viz_mrna_per_mir}: Visualize the number of correlated mRNAs per miRNA

\item \code{viz_neg_cor_fraction}: Fraction of negative correlation events
}}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
A. Kassambara. Statistical Tools for High-throughput Data
  Analysis. http://www.sthda.com
}

