% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_exprs.R
\docType{data}
\name{gene_exprs}
\alias{gene_exprs}
\title{Expression value of miRNA potential targets}
\format{A data frame with 94 rows and 4 columns (sample 1 to 4).}
\usage{
data("gene_exprs")
}
\description{
Expression value of miR-28-5p and miR-768-5p potential targets
}
\examples{
data(gene_exprs)
head(gene_exprs)

}

