% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mirecords.R
\name{get_mirecords}
\alias{get_mirecords}
\alias{get_mirecords_pt}
\alias{get_mirecords_vt}
\title{Get predicted and validated targets for miRNAs using miRecords database}
\usage{
get_mirecords_pt(miRNAs, n = 5, species = "Homo sapiens",
  result_file = "mirna_mirecords_predicted_targets.txt",
  keep_db_names = FALSE, show_progress = TRUE)

get_mirecords_vt(miRNAs, species = "Homo sapiens", version = 4)
}
\arguments{
\item{miRNAs}{a character vector containing miRNA names (e.g.: miRNAs =
c("hsa-miR-16", "hsa-miR-34a))}

\item{n}{integer specifying the minimal number of prediction for each miRNA.
In other words, Only predicted target by at least n/11 database are kept.
Default value is 5.}

\item{species}{species names as available on miRecords database. Possible
values are one of: "Homo sapiens", "Mus musculus", "Rattus norvegicus",
"Caenorhabditis elegans", "Drosophila melanogaster", "Gallus gallus", "Ovis
aries", "Canis familiaris". Default value is "Homo sapiens".}

\item{result_file}{path to result file for holding miRNA predicted/validated
targets. Default values are "mirna_mirecords_predicted_targets.txt".}

\item{keep_db_names}{logical value. If TRUE, keep the name of the 11 databases used for target predictions.}

\item{show_progress}{logical value. If TRUE, a progress bar is shown.}

\item{version}{miRecords version. Default value is 4}
}
\value{
\itemize{ \item \bold{get_mirecords_pt()}: Downloads and returns an object of class MirTarget which is a
data frame containing predicted targets for each miRNA. It saves also the
results as a tab-delimited .txt file, which default name is
"mirna_mirecords_predicted_targets.txt". If this file exists already in your
working directory, then get_mirecords_pt() will simply import it into R.
\item \bold{get_mirecords_vt()}:  Downloads and returns an object of class MirTarget which is a data frame
containing validated targets for each miRNA.

}
}
\description{
miRecords (\url{http://c1.accurascience.com/miRecords/}) is a
  resource for animal miRNA-target interactions. There are two components in
  miRecords. The experimentally Validated Targets (vt) component and the
  Predicted Targets (pt) component which is an integration of predicted miRNA
  targets produced by 11 established miRNA target prediction programs.
  \itemize{

  \item \bold{get_mirecords_pt()}: Recursively retrieve the predicted targets
  for each miRNA. Only predicted targets by at least n/11 database are kept,
  where \bold{n} is user-defined. Internet connexion is required the first
  time you use this function. \item \bold{get_mirecords_vt()}: Download
  validated targets for each miRNA. Internet connexion is required. }
}
\examples{
\donttest{
# Predicted targets
 ptget <- get_mirecords_pt(c("hsa-miR-16", "hsa-miR-155"))
 head(ptget)

# Validated targets
 vtget <- get_mirecords_vt(c("hsa-miR-16", "hsa-miR-155"))
 head(vtget)

}

}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
A. Kassambara. Statistical Tools for High-throughput Data
  Analysis. http://www.sthda.com
}

