% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mirtarbase.R
\name{get_miretarbase}
\alias{get_miretarbase}
\alias{get_mirtarbase_vt}
\title{Get validated targets for miRNAs using miRTarBase database}
\usage{
get_mirtarbase_vt(miRNAs, species = "Homo sapiens", version = 6.1)
}
\arguments{
\item{miRNAs}{a character vector containing miRNA names (e.g.: miRNAs =
c("hsa-miR-16", "hsa-miR-34a))}

\item{species}{of interest as available on miRTarbase (see
http://mirtarbase.mbc.nctu.edu.tw/php/download.php). Allowed values are one
of: 'Arabidopsis thaliana', 'Bombyx mori', 'Bos taurus', 'Caenorhabditis
elegans', 'Drosophila melanogaster', 'Danio rerio', 'Epstein Barr virus',
'Gallus gallus', 'Human cytomegalovirus', 'Homo sapiens', 'Kaposi
sarcoma-associated herpesvirus', 'Mus musculus', 'Ovis aries', 'Oryza
sativa', 'Rattus norvegicus', 'Vesicular stomatitis Indiana virus',
'Xenopus laevis', 'Xenopus tropicalis'.}

\item{version}{the current version of miRTarbase}
}
\value{
\itemize{ \item \bold{get_mirtarbase_vt()}:  Downloads and returns an
  object of class .MirTarget which is a data frame containing validated
  targets for each miRNA. Internet connexion is required.
  }
}
\description{
miRTarBase (\url{http://mirtarbase.mbc.nctu.edu.tw/index.php})
  is an experimentally validated microRNA-target interactions (MTIs)
  database. MTIs are validated experimentally by reporter assay, western
  blot, microarray and next-generation sequencing experiments. The miRTarBase
  provides the most updated collection by comparing with other similar,
  previously developed databases. \itemize{ \item \bold{get_mirtarbase_vt()}:
  Download validated targets for each miRNA using miRTarBase. Internet
  connexion is required. }
}
\examples{
\dontrun{
# 3. Get validated targets for miRNA of interest
 vtget <- get_mirtarbase_vt(c("hsa-miR-16", "hsa-miR-155"))

 head(vtget)
}

}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
A. Kassambara. Statistical Tools for High-throughput Data
  Analysis. http://www.sthda.com
}

