% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{MirMrnaExprs}
\alias{MirMrnaExprs}
\alias{MirMrnaExprs-class}
\title{MirMrnaExprs object and constructor}
\usage{
MirMrnaExprs(data.frame)
}
\arguments{
\item{data.frame}{a data.frame of combine_mir_mrna_exprs()}
}
\value{
a MirMrnaExprs object
}
\description{
MirMrnaExprs extends data.frame. It is used by
\code{\link{combine_mir_mrna_exprs}} to wrap up the results. This constructor
function would not typically be used by "end users".
}

