% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{MirTarget}
\alias{MirTarget}
\alias{MirTarget-class}
\title{MirTarget object and constructor}
\usage{
MirTarget(data.frame)
}
\arguments{
\item{data.frame}{a data frame}
}
\value{
a MirTarget object
}
\description{
MirTarget extends data.frame. It is used to wrap up the results
of predicted and validated target. This constructor
function would not typically be used by "end users".
}

