% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{viz_neg_cor_fraction}
\alias{viz_neg_cor_fraction}
\title{describeIn correlation Fraction of negative correlation events}
\usage{
viz_neg_cor_fraction(object, output = c("plot", "table"))
}
\description{
describeIn correlation Fraction of negative correlation events
}

